backend/requirements-dev.txt
chalice
mysql-connector-python
boto3
